Revert "tmpfs: fix up testing for linux and plan 9"
This reverts commit b4c7f9a1f877b81ef613073732d020127ea64023.
